implement user service  tests

+++ src/test/java/com/example/petfinder/service/UserServiceTests.java
@@ -1,5 +1,7 @@
 package com.example.petfinder.service;
 
+import com.example.petfinder.dto.user.request.PasswordUpdateRequest;
+import com.example.petfinder.dto.user.request.UserUpdateRequest;
 import com.example.petfinder.dto.user.response.UserView;
 import com.example.petfinder.exceptions.PermissionException;
 import com.example.petfinder.mapper.UserMapper;
@@ -18,13 +20,14 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
 import java.time.LocalDate;
-import java.util.Collections;
 import java.util.Optional;
 import java.util.UUID;
 
+import static org.hibernate.validator.internal.util.Contracts.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.*;
 
 @Slf4j
@@ -37,13 +40,15 @@ public class UserServiceTests {
     @Mock
     UserMapper userMapper;
 
+    @Mock
+    PasswordEncoder passwordEncoder;
 
     @InjectMocks
     UserServiceImpl userService;
 
 
     @Test
-    public void testCreateUser_Successful() {
+    public void UserService_CreateUser_Successful() {
         User user = createUser();
 
         when(userRepository.existsByEmail(user.getEmail())).thenReturn(false);
@@ -62,9 +67,9 @@ public class UserServiceTests {
         when(userRepository.findById(user.getId())).thenReturn(Optional.of(user));
         when(userService.getUserById(user.getId())).thenReturn(userView);
 
-        UserView savedUser = userService.getUserById(user.getId());
+        UserView foundUser = userService.getUserById(user.getId());
 
-        Assertions.assertThat(savedUser).isNotNull();
+        Assertions.assertThat(foundUser).isNotNull();
     }
 
     @Test
@@ -81,18 +86,67 @@ public class UserServiceTests {
 
     @Test
     public void UserService_DeleteUserById_ReturnUser() {
-        setUpSecurityContext();
+        User authenticatedUser = getUser();
         User user = createUser();
-        userRepository.save(user);
-        when(userRepository.findAll()).thenReturn(Collections.singletonList(user));
-        userService.deleteUser(user.getId());
+
+        when(userRepository.findUserByEmail(user.getEmail())).thenReturn(Optional.of(user));
+
+        if (authenticatedUser.getEmail().equals(user.getEmail())) {
+            userService.deleteUser(user.getId());
+
+        } else throw new PermissionException("Permission Denied");
+
 
         verify(userRepository).deleteById(user.getId());
     }
 
+    @Test
+    public void UserService_UpdateUserById_ReturnUser() {
+        User updatedUser = createUser();
+        User currentUser = getUser();
+        UserUpdateRequest userUpdateRequest = createUserUpdateRequest();
+        UserView expectedUserView = createUserView();
+
+        when(userRepository.findUserByEmail(currentUser.getEmail())).thenReturn(Optional.of(currentUser));
+        when(userRepository.save(any(User.class))).thenReturn(currentUser);
+        when(userMapper.userToView(any(User.class))).thenReturn(expectedUserView);
+
+        if (!updatedUser.getUsername().equals(currentUser.getUsername())) {
+
+            throw new PermissionException("Permission denied");
+        }
+
+        UserView result = userService.updateUser(currentUser.getId(), userUpdateRequest);
+
+        assertNotNull(result);
+        assertEquals(expectedUserView, result);
+
+        verify(userRepository, times(1)).save(any(User.class));
+        verify(userMapper, times(1)).userToView(any(User.class));
+    }
+
+    @Test
+    public void UserService_UpdatePassword_ReturnSuccess() {
+        User user = getUser();
+        PasswordUpdateRequest passwordUpdateRequest = createPasswordUpdateRequest();
+
+        when(userRepository.findUserByEmail(anyString())).thenReturn(Optional.ofNullable(user));
+        when(passwordEncoder.matches(anyString(), anyString())).thenReturn(true);
+
+        userService.updatePassword(passwordUpdateRequest);
+
+        verify(passwordEncoder, times(1)).encode(passwordUpdateRequest.newPassword());
+    }
+
+    private PasswordUpdateRequest createPasswordUpdateRequest() {
+        return PasswordUpdateRequest.builder()
+                .oldPassword("liza230)")
+                .newPassword("liza23430))")
+                .build();
+    }
+
     private User createUser() {
         UUID userId = UUID.randomUUID();
-
         return User.builder()
                 .id(userId)
                 .email("elizabeth@gmail.com")
@@ -107,6 +161,15 @@ public class UserServiceTests {
                 .build();
     }
 
+    private UserUpdateRequest createUserUpdateRequest() {
+        return UserUpdateRequest.builder()
+                .firstName("Liza")
+                .lastName("Doe")
+                .biography("something")
+                .sex(Sex.MALE)
+                .build();
+    }
+
     private UserView createUserView() {
         return UserView.builder()
                 .email("elizabeth@gmail.com")
@@ -116,39 +179,16 @@ public class UserServiceTests {
                 .build();
     }
 
-    public User getCurrentUser() {
-        var username = SecurityContextHolder.getContext().getAuthentication().getName();
-        return getByUsername(username);
-    }
 
-    public User getByUsername(String username) {
-        return userRepository.findUserByEmail(username).orElseThrow(() -> new UsernameNotFoundException("User not found"));
-
-    }
-
-    private void checkPermission(UUID userId) {
-
-        if (!getCurrentUser().getId().equals(userId)) {
-            throw new PermissionException();
-        }
-    }
-
-    private void setCurrentUser(User user) {
-        Authentication authentication = mock(Authentication.class);
-        when(authentication.getName()).thenReturn(user.getEmail());
-        SecurityContext securityContext = mock(SecurityContext.class);
-        when(securityContext.getAuthentication()).thenReturn(authentication);
-        SecurityContextHolder.setContext(securityContext);
-    }
-
-    private void setUpSecurityContext() {
+    private User getUser() {
         SecurityContext securityContext = mock(SecurityContext.class);
         Authentication authentication = mock(Authentication.class);
 
-        when(authentication.getName()).thenReturn("testUser");
+        when(authentication.getName()).thenReturn("elizabeth@gmail.com");
         when(securityContext.getAuthentication()).thenReturn(authentication);
 
         SecurityContextHolder.setContext(securityContext);
-    }
 
+        return createUser();
+    }
 }
